Refactor: move swap button component to swap feature
tmp
