+++ package.json
@@ -61,9 +61,6 @@
     "@safe-global/safe-apps-sdk": "^9.0.0-next.1",
     "@safe-global/safe-deployments": "^1.35.0",
     "@safe-global/safe-gateway-typescript-sdk": "3.21.1",
-    "@safe-global/safe-core-sdk": "^3.3.5",
-    "@safe-global/safe-core-sdk-utils": "^1.7.4",
-    "@safe-global/safe-ethers-lib": "^1.9.4",
     "@safe-global/safe-modules-deployments": "^1.2.0",
     "@sentry/react": "^7.91.0",
     "@spindl-xyz/attribution-lite": "^1.4.0",
